some inting

+++ spammo/payment.py
@@ -57,10 +57,10 @@ def _pay_or_charge(user, amount, note):
     	# print "over 60 cents"
     	for x in range(0,60):
     		if x < 60:
-    			realAmount = (pennies / 60) / 100.00
+    			realAmount = int(pennies / 60) / 100.00
     		else:
-    			leftoverpennies = pennies % 60
-    			realAmount = ((pennies / 60) + leftoverpennies) / 100.00
+    			leftoverpennies = int(pennies % 60)
+    			realAmount = (int(pennies / 60) + leftoverpennies) / 100.00
     		# print realAmount
     		payments.append(realAmount)
     else:
@@ -107,7 +107,7 @@ def _pay_or_charge(user, amount, note):
 	    # note = payment['note']
 	    print(str(i) + '. Successfully {payment_action} {user} ${amount:.2f} for "{note}"'
 	           .format(**locals()))
-	    
+
     print("You just sent a total of $"+amount+" in "+len(payments)+" payments of $"+payments[0]". I hope you're happy.")
